requirements.txt
pandas
pyyaml
requests
beautifulsoup4
lxml
pdfminer.six